Tidying up the moved models
