+++ Analyser/src/Models/Helpers.js
@@ -1,4 +1,7 @@
 import Config from "../Config";
+import { ConcolicValue } from "../Values/WrappedValue";
+
+const map = Array.prototype.map;
 
 export default function(state, ctx) {
 
@@ -104,6 +107,63 @@ export default function(state, ctx) {
 		};
 	}
 
+
+	/**
+	 * In JavaScript slice and substr can be given a negative index to indicate addressing from the end of the array
+	 * We need to rewrite the SMT to handle these cases
+	 */
+	function substringHandleNegativeLengths(base_s, index_s) {
+
+		//Index s is negative to adding will get us to the right start
+		const newIndex = ctx.mkAdd(base_s.getLength(), index_s);
+
+		//Bound the minimum index by 0
+		const aboveMin = ctx.mkGe(newIndex, ctx.mkIntVal(0));
+		const indexOrZero = ctx.mkIte(aboveMin, newIndex, ctx.mkIntVal(0));
+
+		return ctx.mkIte(ctx.mkGe(index_s, ctx.mkIntVal(0)), index_s, indexOrZero);
+	}
+
+	function substringHelper(base, args, result) {
+		state.stats.seen("Symbolic Substrings");
+
+		const target = state.asSymbolic(base);
+
+		//The start offset is either the argument of str.len - the arguments
+		let start_off = ctx.mkRealToInt(state.asSymbolic(args[0]));
+		start_off = substringHandleNegativeLengths(target, start_off);
+
+		//Length defaults to the entire string if not specified
+		let len;
+		const maxLength = ctx.mkSub(target.getLength(), start_off);
+
+		if (args[1]) {
+			len = state.asSymbolic(args[1]);
+			len = ctx.mkRealToInt(len);
+
+			//If the length is user-specified bound the length of the substring by the maximum size of the string ("123".slice(0, 8) === "123")
+			const exceedMax = ctx.mkGe(
+					ctx.mkAdd(start_off, len),
+					target.getLength()
+					);
+
+			len = ctx.mkIte(exceedMax, maxLength, len);
+		} else {
+			len = maxLength;
+		}
+
+		//If the start index is greater than or equal to the length of the string the empty string is returned
+		const substr_s = ctx.mkSeqSubstr(target, start_off, len);
+		const empty_s = ctx.mkString("");
+		const result_s = ctx.mkIte(
+				ctx.mkGe(start_off, target.getLength()),
+				empty_s,
+				substr_s
+				);
+
+		return new ConcolicValue(result, result_s);
+	}
+
 	return {
 		symbolicHook: symbolicHook,
 		DoesntMatch: DoesntMatch,
@@ -111,6 +171,7 @@ export default function(state, ctx) {
 		ConcretizeIfNative: ConcretizeIfNative,
 		coerceToString: coerceToString,
 		symbolicHookRe: symbolicHookRe,
-		NoOp: NoOp
+		NoOp: NoOp,
+		substring: substringHelper
 	};
 }
