+++ Analyser/src/Models/RegexModels.js
@@ -1,6 +1,12 @@
+import Z3 from "z3javascript";
+import { ConcolicValue } from "../Values/WrappedValue";
+import Config from "../Config";
+import Log from "../Utilities/Log";
+
 export default function(state, ctx, model, helpers) {
 
 	const symbolicHookRe = helpers.symbolicHookRe;
+	const coerceToString = helpers.coerceToString;
 
 	function EnableCaptures(regex, real, string_s) {
 
@@ -183,62 +189,6 @@ trueCheck: [NotMatch, CheckFixed],
 				);
 
 		return new ConcolicValue(result, search_in_re);
-	}
-
-	/**
-	 * In JavaScript slice and substr can be given a negative index to indicate addressing from the end of the array
-	 * We need to rewrite the SMT to handle these cases
-	 */
-	function substringHandleNegativeLengths(base_s, index_s) {
-
-		//Index s is negative to adding will get us to the right start
-		const newIndex = ctx.mkAdd(base_s.getLength(), index_s);
-
-		//Bound the minimum index by 0
-		const aboveMin = ctx.mkGe(newIndex, ctx.mkIntVal(0));
-		const indexOrZero = ctx.mkIte(aboveMin, newIndex, ctx.mkIntVal(0));
-
-		return ctx.mkIte(ctx.mkGe(index_s, ctx.mkIntVal(0)), index_s, indexOrZero);
-	}
-
-	function substringHelper(base, args, result) {
-		state.stats.seen("Symbolic Substrings");
-
-		const target = state.asSymbolic(base);
-
-		//The start offset is either the argument of str.len - the arguments
-		let start_off = ctx.mkRealToInt(state.asSymbolic(args[0]));
-		start_off = substringHandleNegativeLengths(target, start_off);
-
-		//Length defaults to the entire string if not specified
-		let len;
-		const maxLength = ctx.mkSub(target.getLength(), start_off);
-
-		if (args[1]) {
-			len = state.asSymbolic(args[1]);
-			len = ctx.mkRealToInt(len);
-
-			//If the length is user-specified bound the length of the substring by the maximum size of the string ("123".slice(0, 8) === "123")
-			const exceedMax = ctx.mkGe(
-					ctx.mkAdd(start_off, len),
-					target.getLength()
-					);
-
-			len = ctx.mkIte(exceedMax, maxLength, len);
-		} else {
-			len = maxLength;
-		}
-
-		//If the start index is greater than or equal to the length of the string the empty string is returned
-		const substr_s = ctx.mkSeqSubstr(target, start_off, len);
-		const empty_s = ctx.mkString("");
-		const result_s = ctx.mkIte(
-				ctx.mkGe(start_off, target.getLength()),
-				empty_s,
-				substr_s
-				);
-
-		return new ConcolicValue(result, result_s);
 	}
 
 	/**
